+++ Changes.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Added
+- Forgot to include the working example of HTTP::Tiny + Parataxis with the last dist.
+
 ## [v0.0.9] - 2026-02-21
 
 Asynchronous HTTP::Tiny is basically a semi-automatic footgun.